+++ src/kademliaRPC/RPC/ping.go
@@ -1,0 +1,1 @@
+package rpc
