+++ src/kademliaRPC/RPC/store.go
@@ -1,0 +1,1 @@
+package rpc
